requirements.txt
# =====================================================================
# requirements.txt — Phase 1 (Classification 2D ISIC 2019)
# Tested with Python 3.11 + CUDA 12.1
# =====================================================================

# ---- Deep Learning Core -----------------------------------------
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0

# ---- Vision models (architectures) ------------------------------
timm==1.0.3                     # ResNet, DeiT, ViT, Swin
transformers==4.40.0            # HuggingFace (ViT, DINOv2)

# ---- XAI — classical methods ------------------------------------
captum==0.7.0                   # IG, LRP, SHAP, Occlusion (Meta AI)
grad-cam==1.5.0                 # Grad-CAM, Grad-CAM++
shap==0.45.0
lime==0.2.0.1

# ---- XAI — mechanistic (Phase 1 contributions C1, C2) -----------
sae-lens==3.0.0                 # Sparse Autoencoders
transformer-lens==1.17.0        # Activation Patching, Circuit Discovery

# ---- Configuration & experiment tracking ------------------------
hydra-core==1.3.2               # Hierarchical YAML configs
omegaconf==2.3.0
wandb==0.17.0                   # Experiment tracking

# ---- Data processing --------------------------------------------
albumentations==1.4.0           # Augmentations
opencv-python-headless==4.9.0.80
Pillow==10.3.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
scikit-image==0.22.0

# ---- Visualization ----------------------------------------------
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.20.0

# ---- Notebooks --------------------------------------------------
jupyter==1.0.0
jupyterlab==4.1.6
ipywidgets==8.1.2
nbformat==5.10.4

# ---- Utilities --------------------------------------------------
loguru==0.7.2                   # Better logging
tqdm==4.66.2
python-dotenv==1.0.1
rich==13.7.1                    # Pretty terminal output

# ---- Code quality (dev) -----------------------------------------
ruff==0.4.1
black==24.4.0
mypy==1.9.0
pytest>=8.1.2,<9.0.0
pytest-cov==5.0.0
pre-commit==3.7.0

# ---- Reproducibility --------------------------------------------
codecarbon==2.4.1               # Track carbon emissions
